+++ src/components/Hero/Hero.tsx
@@ -204,7 +204,7 @@ export const Hero: ChakraComponent<
         backgroundImageStyle = backgroundImageSrc
           ? {
               bgColor: defaultBackgroundColor,
-              backgroundImage: `url(${backgroundImageSrc})`,
+              backgroundImage: `/**/url("${backgroundImageSrc}")`,
             }
           : {};
       } else if (heroType === "secondary") {
@@ -236,8 +236,8 @@ export const Hero: ChakraComponent<
                   ? "saturation"
                   : null,
                 backgroundImage: isDarkBackgroundImage
-                  ? `linear-gradient(black, black), url(${backgroundImageSrc})`
-                  : `url(${backgroundImageSrc})`,
+                  ? `/**/linear-gradient(black, black), url("${backgroundImageSrc}")`
+                  : `/**/url("${backgroundImageSrc}")`,
                 backgroundPosition: "center",
                 backgroundSize: "cover",
                 opacity: isDarkBackgroundImage ? "0.4" : "1.0",
@@ -306,7 +306,9 @@ export const Hero: ChakraComponent<
             <Box
               __css={{
                 ...styles.imgWrapper,
-                backgroundImage: `url(${imageProps.src})`,
+              }}
+              style={{
+                backgroundImage: `/**/url("${imageProps.src}")`,
               }}
             >
               {imageToRender}
@@ -322,9 +324,11 @@ export const Hero: ChakraComponent<
               <Box
                 __css={{
                   ...styles.imgWrapper,
+                }}
+                style={{
                   backgroundImage:
                     heroType === "fiftyFifty"
-                      ? `url(${imageProps.src})`
+                      ? `/**/url("${imageProps.src}")`
                       : undefined,
                 }}
               >
@@ -344,12 +348,11 @@ export const Hero: ChakraComponent<
         <Box
           data-testid="hero"
           data-responsive-background-image
-          style={backgroundImageSrc ? backgroundImageStyle : undefined}
           ref={ref}
           __css={{
             ...styles.base,
-            ...backgroundImageStyle,
           }}
+          style={backgroundImageSrc ? backgroundImageStyle : undefined}
         >
           <Box
             data-testid="hero-content"